fix(video-editor): drop direct service import from captions screen
The ui_service_boundary ratchet rejects new UI-layer imports of
lib/services. Move the generation outcome types into models, default
the production service inside CaptionsEditorCubit (mirroring
AudioTimingCubit), and swap the screen's test seam from a service
override to a cubit override.
